requirements.txt
datasets>=2.16.0
qdrant-client>=1.7.0
sentence-transformers>=2.5.0
torch>=2.2.0
groq>=0.4.2
python-dotenv>=1.0.0
tqdm>=4.66.0
pandas>=2.2.0
numpy>=1.26.0 